cacheable - fix: moving isKeyvInstance to utils
